feat : 우숭자를 보여주는 기능 생성
