--- src/components/showResults.js
@@ -1,5 +1,0 @@
-import { showResultContainer } from '../utils/showTags.js';
-
-export default function showResults(racingCars) {
-  showResultContainer();
-}
